requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0